test: consolidate function-wrapper tests into parser_ir_advanced_test.go
Move the contents of parser_ir_function_wrapper_test.go into the existing
parser_ir_advanced_test.go and delete the standalone file. This continues
the trend of merging narrowly-scoped test files back into the broader
advanced-IR suite rather than growing the flat top-level file count.

The strPtr / findUsageWithFunction / flattenColumnUsage helpers live with
the function-wrapper tests now and are only referenced from those tests,
so consolidation is a no-op for compile + go test.
